+++ src/server/Player.h
@@ -20,7 +20,7 @@ class Player {
 
         std::string get_match_name();
 
-        void send(std::vector<uint8_t>& msg);
+        void send(std::vector<int32_t>& msg);
 
         void send(std::string& msg);
